feat(flasbackWebUI): [6] split the web interface in a dedicated structure
